use float instead of double

+++ src/PNS.h
@@ -30,7 +30,7 @@ public:
         var dn_th = 1;
 
       bool extended = false;
-      double heuristic_value;
+      float heuristic_value;
       unsigned int child_num;
